Add db filler data sql files

+++ Backend/src/main/java/com/codecool/backend/controller/dto/NewTransactionDto.java
@@ -2,5 +2,5 @@ package com.codecool.backend.controller.dto;
 
 import com.codecool.backend.model.Member;
 
-public record NewTransactionDto(String name, int categoryId, int amount, int memberId) {
+public record NewTransactionDto(String name, Long categoryId, int amount, int memberId) {
 }
